refactor: move NewRandomMigration in libmigration package
