+++ tests/storage/hotplug.go
@@ -1249,7 +1249,7 @@ var _ = SIGDescribe("Hotplug", func() {
 					}
 					Expect(sourceAttachmentPods).To(HaveLen(1))
 
-					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
+					migration := libmigration.NewRandomMigration(vmi.Name, vmi.Namespace)
 					migration = libmigration.RunMigrationAndExpectToCompleteWithDefaultTimeout(virtClient, migration)
 					libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 					By("Verifying the volume is still accessible and usable")
