+++ tests/vmi_tpm_test.go
@@ -76,7 +76,7 @@ var _ = Describe("[sig-compute]vTPM", decorators.SigCompute, decorators.Requires
 
 			By("Migrating the VMI")
 			checks.SkipIfMigrationIsNotPossible()
-			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
+			migration := libmigration.NewRandomMigration(vmi.Name, vmi.Namespace)
 			libmigration.RunMigrationAndExpectToCompleteWithDefaultTimeout(virtClient, migration)
 
 			By("Ensuring the TPM is still functional and its state carried over")
